feat: use a CGO-less sqlite database alternative ncruces/go-sqlite3
