+++ .github/workflows/build_and_deploy.yaml
@@ -109,3 +109,19 @@ jobs:
           tags: |
             ${{ env.REGISTRY }}/${{ env.REPOSITORY }}/frontend:latest
             ${{ env.REGISTRY }}/${{ env.REPOSITORY }}/frontend:${{ env.GIT_HASH }}
+
+      # ---------------------------------------------------------------------
+
+      - name: "Build :: litestream"
+        uses: docker/build-push-action@v4
+        with:
+          context: ./litestream
+          file: litestream.Dockerfile
+          build-args: |
+            buildtime_variable_version=${{ env.BUILD_ID }}
+            buildtime_variable_timestamp=${{ env.BUILD_DATE }}
+            buildtime_variable_commit=${{ env.GIT_HASH }}
+          push: true
+          tags: |
+            ${{ env.REGISTRY }}/${{ env.REPOSITORY }}/litestream:latest
+            ${{ env.REGISTRY }}/${{ env.REPOSITORY }}/litestream:${{ env.GIT_HASH }}
