+++ bookmarks.Dockerfile
@@ -6,13 +6,11 @@ ARG buildtime_variable_version=1.0.0
 ARG buildtime_variable_timestamp=20220101
 ARG buildtime_variable_commit=dev
 ARG buildtime_variable_arch=amd64
-ARG buildtime_variable_litestream_ver=v0.3.13/litestream-v0.3.13-linux-amd64.tar.gz
 
 ENV VERSION=${buildtime_variable_version}
 ENV TSTAMP=${buildtime_variable_timestamp}
 ENV COMMIT=${buildtime_variable_commit}
 ENV ARCH=${buildtime_variable_arch}
-ENV LSV=${buildtime_variable_litestream_ver}
 
 WORKDIR /backend-build
 COPY ./go.mod ./
@@ -23,42 +21,41 @@ COPY ./internal/common  ./internal/common
 COPY ./pkg ./pkg
 COPY ./assets ./assets
 
-# necessary to build sqlite3
-RUN apk add build-base
-
 ##
 ## go build
-RUN GOOS=linux GOARCH=${ARCH} go build -ldflags="-s -w -X main.Version=${TSTAMP} -X main.Build=${COMMIT}" -o bookmarks.api ./cmd/bookmarks/server/main.go
-
-##
-## include litestream into the image and use the litestream replication capabilities
-ADD https://github.com/benbjohnson/litestream/releases/download/${LSV} /backend-build/litestream.tar.gz
-RUN tar -C /backend-build -xzf /backend-build/litestream.tar.gz
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=${ARCH} go build -ldflags="-s -w -X main.Version=${TSTAMP} -X main.Build=${COMMIT}" -o bookmarks.api ./cmd/bookmarks/server/main.go
 
 ## --------------------------------------------------------------------------
 
 ## runtime
 ## --------------------------------------------------------------------------
-FROM alpine:latest
+FROM alpine:3
+
+ARG buildtime_variable_username="containeruser"
+ARG buildtime_variable_groupname="containergroup"
+ARG buildtime_variable_uid="65532"
+ARG buildtime_variable_gid="65532"
+ARG buildtime_variable_port="3000"
+
 LABEL author="henrik@binggl.net"
 WORKDIR /opt/bookmarks
-RUN mkdir -p /opt/litestream && mkdir -p /opt/bookmarks/etc && mkdir -p /opt/bookmarks/logs && mkdir -p /opt/bookmarks/uploads && mkdir -p /opt/bookmarks/db && mkdir -p /opt/bookmarks/assets
-EXPOSE 3000
-
-RUN apk add bash
+RUN mkdir -p /opt/bookmarks/etc && mkdir -p /opt/bookmarks/logs && mkdir -p /opt/bookmarks/uploads && mkdir -p /opt/bookmarks/db && mkdir -p /opt/bookmarks/assets
 
 # Do not run as root user
 ## alpine specific user/group creation
-RUN addgroup -g 1000 -S bookmarks && \
-    adduser -u 1000 -S bookmarks -G bookmarks
-COPY --chown=1000:1000 --from=BACKEND-BUILD /backend-build/bookmarks.api /opt/bookmarks
-COPY --chown=1000:1000 --from=BACKEND-BUILD /backend-build/litestream /opt/litestream
-COPY --chown=1000:1000 --from=BACKEND-BUILD /backend-build/assets /opt/bookmarks/assets
-COPY --chown=1000:1000 ./litestream/run_litestream.sh /opt/bookmarks
-RUN chown bookmarks:bookmarks /opt/bookmarks/etc \
-    && chown bookmarks:bookmarks /opt/bookmarks/logs \
-    && chown bookmarks:bookmarks /opt/bookmarks/uploads \
-    && chown bookmarks:bookmarks /opt/bookmarks/db
-USER bookmarks
+RUN addgroup -g ${buildtime_variable_gid} -S ${buildtime_variable_groupname} && \
+    adduser -u ${buildtime_variable_uid} -S ${buildtime_variable_username} -G ${buildtime_variable_groupname} -H -h /opt/bookmarks
 
-CMD [ "/opt/bookmarks/run_litestream.sh" ]
+COPY --chown=${buildtime_variable_uid}:${buildtime_variable_gid} --from=BACKEND-BUILD /backend-build/bookmarks.api /opt/bookmarks
+COPY --chown=${buildtime_variable_uid}:${buildtime_variable_gid} --from=BACKEND-BUILD /backend-build/assets /opt/bookmarks/assets
+
+RUN chown ${buildtime_variable_uid}:${buildtime_variable_gid} /opt/bookmarks/etc \
+    && chown ${buildtime_variable_uid}:${buildtime_variable_gid} /opt/bookmarks/logs \
+    && chown ${buildtime_variable_uid}:${buildtime_variable_gid} /opt/bookmarks/uploads \
+    && chown ${buildtime_variable_uid}:${buildtime_variable_gid} /opt/bookmarks/db
+
+USER ${buildtime_variable_username}
+
+EXPOSE ${buildtime_variable_port}
+
+CMD [ "/opt/bookmarks/bookmarks.api", "--basepath=/opt/bookmarks", "--hostname=0.0.0.0" ]
