+++ internal/core/app/store/db_repository_test.go
@@ -17,8 +17,8 @@ func repo(t *testing.T) (store.Repository, *sql.DB) {
 	var (
 		err error
 	)
-	params := make([]persistence.SqliteParam, 0)
-	con, err := persistence.CreateGormSqliteCon(":memory:", params)
+	dbCon := persistence.MustCreateSqliteConn(":memory:")
+	con, err := persistence.CreateGormSqliteCon(dbCon)
 	if err != nil {
 		t.Fatalf("cannot create database connection: %v", err)
 	}
