+++ mydms.Dockerfile
@@ -6,13 +6,11 @@ ARG buildtime_variable_version=3.0.0
 ARG buildtime_variable_timestamp=20220101
 ARG buildtime_variable_commit=dev
 ARG buildtime_variable_arch=amd64
-ARG buildtime_variable_litestream_ver=v0.3.13/litestream-v0.3.13-linux-amd64.tar.gz
 
 ENV VERSION=${buildtime_variable_version}
 ENV TSTAMP=${buildtime_variable_timestamp}
 ENV COMMIT=${buildtime_variable_commit}
 ENV ARCH=${buildtime_variable_arch}
-ENV LSV=${buildtime_variable_litestream_ver}
 
 WORKDIR /backend-build
 COPY ./cmd ./cmd
@@ -21,47 +19,43 @@ COPY ./go.sum ./
 COPY ./internal/mydms ./internal/mydms
 COPY ./internal/common  ./internal/common
 COPY ./pkg ./pkg
-COPY ./tools ./tools
 COPY ./assets ./assets
-
-# necessary to build sqlite3
-RUN apk add build-base
 
 ##
 ## go build
-RUN GOOS=linux GOARCH=${ARCH} go build -ldflags="-w -s -X main.Version=${TSTAMP} -X main.Build=${COMMIT}" -o mydms.api ./cmd/mydms/server/*.go
-
-##
-## include litestream into the image and use the litestream replication capabilities
-ADD https://github.com/benbjohnson/litestream/releases/download/${LSV} /backend-build/litestream.tar.gz
-RUN tar -C /backend-build -xzf /backend-build/litestream.tar.gz
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=${ARCH} go build -ldflags="-w -s -X main.Version=${TSTAMP} -X main.Build=${COMMIT}" -o mydms.api ./cmd/mydms/server/*.go
 
 ## --------------------------------------------------------------------------
 
 ## runtime
 ## --------------------------------------------------------------------------
-FROM alpine:latest
+FROM alpine:3
+
+ARG buildtime_variable_username="containeruser"
+ARG buildtime_variable_groupname="containergroup"
+ARG buildtime_variable_uid="65532"
+ARG buildtime_variable_gid="65532"
+ARG buildtime_variable_port="3000"
+
 LABEL author="henrik@binggl.net"
 WORKDIR /opt/mydms
-RUN mkdir -p /opt/litestream && mkdir -p /opt/mydms/uploads && mkdir -p /opt/mydms/etc && mkdir -p /opt/mydms/logs && mkdir -p /opt/mydms/db && mkdir -p /opt/mydms/assets
-EXPOSE 3000
-
-RUN apk add bash
+RUN mkdir -p /opt/mydms/uploads && mkdir -p /opt/mydms/etc && mkdir -p /opt/mydms/logs && mkdir -p /opt/mydms/db && mkdir -p /opt/mydms/assets
 
 # Do not run as root user
 ## alpine specific user/group creation
-RUN addgroup -g 1000 -S mydms && \
-    adduser -u 1000 -S mydms -G mydms
+RUN addgroup -g ${buildtime_variable_gid} -S ${buildtime_variable_groupname} && \
+    adduser -u ${buildtime_variable_uid} -S ${buildtime_variable_username} -G ${buildtime_variable_groupname} -H -h /opt/mydms
 
-COPY --chown=1000:1000 --from=BACKEND-BUILD /backend-build/mydms.api /opt/mydms
-COPY --chown=1000:1000 --from=BACKEND-BUILD /backend-build/assets /opt/mydms/assets
-COPY --chown=1000:1000 --from=BACKEND-BUILD /backend-build/litestream /opt/litestream
-COPY --chown=1000:1000 ./litestream/run_litestream.sh /opt/mydms
-RUN chown mydms:mydms /opt/mydms/etc \
-    && chown mydms:mydms /opt/mydms/logs \
-    &&  chown mydms:mydms /opt/mydms/uploads \
-    &&  chown mydms:mydms /opt/mydms/db
+COPY --chown=${buildtime_variable_uid}:${buildtime_variable_gid} --from=BACKEND-BUILD /backend-build/mydms.api /opt/mydms
+COPY --chown=${buildtime_variable_uid}:${buildtime_variable_gid} --from=BACKEND-BUILD /backend-build/assets /opt/mydms/assets
 
-USER mydms
+RUN chown ${buildtime_variable_uid}:${buildtime_variable_gid} /opt/mydms/etc \
+    && chown ${buildtime_variable_uid}:${buildtime_variable_gid} /opt/mydms/logs \
+    &&  chown ${buildtime_variable_uid}:${buildtime_variable_gid} /opt/mydms/uploads \
+    &&  chown ${buildtime_variable_uid}:${buildtime_variable_gid} /opt/mydms/db
 
-CMD [ "/opt/mydms/run_litestream.sh" ]
+USER ${buildtime_variable_username}
+
+EXPOSE ${buildtime_variable_port}
+
+CMD [ "/opt/mydms/mydms.api", "--basepath=/opt/mydms", "--hostname=0.0.0.0" ]
