--- testdata/integration/pyproject.toml
@@ -1,10 +1,0 @@
-[project]
-name = "monorepo_integration"
-version = "0.1.0"
-description = "Using playwright to test the UI of the monorepo"
-readme = "README.md"
-requires-python = ">=3.12"
-dependencies = [
-    "playwright>=1.49.1",
-    "pytest-playwright>=0.6.2",
-]
