+++ testdata/sqlite/.env.example
@@ -1,0 +1,5 @@
+LITESTREAM_VERSION=v0.3.13/litestream-v0.3.13-linux-amd64.tar.gz
+CORE_REPLICA_URL=s3://<bucket-name>/__MODE__/core.db
+BOOKMARKS_REPLICA_URL=s3://<bucket-name>/__MODE__/bookmarks.db
+MYDMS_REPLICA_URL=s3://<bucket-name>/__MODE__/mydms.db
+
